requirements.txt
neo4j>=4.4.0
python-dotenv>=0.19.0
pydantic>=1.8.2